re-organized utils
