+++ src/utils/point.js
@@ -1,10 +1,5 @@
 import dayjs from 'dayjs';
 
-export const RenderPosition = {
-  AFTER_BEGIN: 'afterbegin',
-  BEFORE_END: 'beforeend',
-};
-
 export const ModeForm = {
   NEW: 'new',
   EDIT: 'edit',
@@ -13,31 +8,6 @@ export const ModeForm = {
 export const ButtonAction = {
   DELETE: 'Delete',
   CANCEL: 'Cancel',
-};
-
-export const getRandomInteger = (a = 0, b = 1) => {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-  return Math.floor(lower + Math.random() * (upper - lower + 1));
-};
-
-export const createElement = (template) => {
-  const newElement = document.createElement('div');
-  newElement.innerHTML = template;
-
-  return newElement.firstChild;
-};
-
-
-export const renderElement = (container, element, place) => {
-  switch (place) {
-    case RenderPosition.AFTER_BEGIN:
-      container.prepend(element);
-      break;
-    case RenderPosition.BEFORE_END:
-      container.append(element);
-      break;
-  }
 };
 
 export const getFullRout = (points) => {
